feat: 의견 삭제 시 이미지 삭제 기능 구현 (#80)

+++ src/main/java/com/jisungin/application/talkroom/TalkRoomService.java
@@ -159,7 +159,7 @@ public class TalkRoomService {
         commentRepository.findByTalkRoom(talkRoom).ifPresent(commentRepository::delete);
 
         List<TalkRoomImage> images = talkRoomImageRepository.findByTalkRoom(talkRoom);
-        if (images != null) {
+        if (images != null && !images.isEmpty()) {
             talkRoomImageRepository.deleteAll(images);
         }
         talkRoomRoleRepository.deleteAllByTalkRoom(talkRoom);
